safari-600.7-branch/LayoutTests/fast/regions/regions-widows-and-orphans-expected.txt

Testing widows and orphans. Any green lines should be at the bottom of regions, and any red lines should be at the top of regions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test1 Block 1 Line 1 is correct.
PASS test1 Block 2 Line 5 is correct.
PASS test1 Block 4 Line 2 is correct.
PASS test2 Block 1 Line 1 is correct.
PASS test2 Block 2 Line 1 is correct.
PASS test3 Block 1 Line 1 is correct.
PASS test3 Block 2 Line 5 is correct.
PASS test4 Block 1 Line 1 is correct.
PASS test4 Block 2 Line 1 is correct.
PASS test4 Block 3 Line 3 is correct.
PASS test5 Block 1 Line 1 is correct.
PASS test5 Block 2 Line 3 is correct.
PASS test6 Block 1 Line 1 is correct.
PASS test6 Block 2 Line 1 is correct.
PASS test6 Block 3 Line 1 is correct.
PASS successfullyParsed is true

TEST COMPLETE
